ratans doc fixes
