+++ example/script/Cargo.toml
@@ -13,6 +13,7 @@ tokio.workspace = true
 
 clap = { version = "4.0", features = ["derive"] }
 sp1-sdk = "2.0.0"
+hex = "0.4.3"
 
 [build-dependencies]
 sp1-build = "2.0.0"
