+++ verifier/Cargo.toml
@@ -20,7 +20,9 @@ ark-serialize = "0.4.2"
 ark-ff = "0.4.2"
 groth16-solana = { git = "https://github.com/sp1-patches/groth16-solana", branch = "patch-v0.0.3" }
 thiserror = "1.0.63"
+hex = "0.4.3"
 
 [dev-dependencies]
 sp1-sdk = { version = "2.0.0", default-features = false}
+hex-literal = "0.3.1"
 num-traits = { version = "0.2.19" }
