+++ verifier/src/test.rs
@@ -29,10 +29,12 @@ fn test_verify_from_sp1() {
     padded_vkey_hash.extend_from_slice(&vkey_hash);
     let vkey_hash = padded_vkey_hash;
 
+    let sp1_vkey_hash = format!("0x{}", hex::encode(vkey_hash));
+
     assert!(verify_proof(
         &proof_bytes,
         &sp1_public_inputs,
-        &vkey_hash[..32].try_into().unwrap(),
+        &sp1_vkey_hash,
         &GROTH16_VK_2_0_0_BYTES
     )
     .is_ok());
@@ -40,7 +42,7 @@ fn test_verify_from_sp1() {
 
 #[test]
 fn test_hash_public_inputs_() {
-    use crate::hash_public_inputs;
+    use crate::utils::hash_public_inputs;
 
     // Read the serialized SP1ProofWithPublicValues from the file.
     let sp1_proof_with_public_values_file = "../proofs/fibonacci_proof.bin";
@@ -59,5 +61,17 @@ fn test_hash_public_inputs_() {
     assert_eq!(
         committed_values_digest,
         hash_public_inputs(&sp1_proof_with_public_values.public_values.to_vec())
+    );
+}
+
+#[test]
+fn test_decode_sp1_vkey_hash() {
+    use crate::utils::decode_sp1_vkey_hash;
+
+    let sp1_vkey_hash = "0x0054c0e58911dd8b993c6d8f249aa50a2e523114ec4b7ef9dd355c5f6bfbf3ce";
+    let decoded_sp1_vkey_hash = decode_sp1_vkey_hash(sp1_vkey_hash).unwrap();
+    assert_eq!(
+        decoded_sp1_vkey_hash,
+        hex_literal::hex!("0054c0e58911dd8b993c6d8f249aa50a2e523114ec4b7ef9dd355c5f6bfbf3ce")
     );
 }
